[T] dialog: move dialog to components
